requirements.txt
PyYAML>=5.4.1
kubernetes>=17.17.0
click>=8.0.1